feat(agent-gui): surface standalone decision toasts
